+++ jwt_allauth/utils.py
@@ -1,9 +1,11 @@
 import warnings
+from datetime import timedelta
 from importlib import import_module
 from typing import Any, Dict, Optional
 
 from allauth.account.adapter import get_adapter
 from allauth.account.models import EmailAddress
+from django.core.exceptions import ImproperlyConfigured
 from django.contrib.auth import get_user_model
 from django.utils.decorators import method_decorator
 from django.views.decorators.debug import sensitive_post_parameters
@@ -14,7 +16,7 @@ from rest_framework import status
 from rest_framework.response import Response
 from rest_framework_simplejwt.exceptions import InvalidToken
 
-from jwt_allauth.constants import TEMPLATE_PATHS, REFRESH_TOKEN_COOKIE
+from jwt_allauth.constants import TEMPLATE_PATHS, REFRESH_TOKEN_COOKIE, DEFAULT_SESSION_LIFETIME
 from jwt_allauth.exceptions import NotVerifiedEmail, IncorrectCredentials
 
 string_types = (str,)
@@ -36,6 +38,24 @@ def _get_cookie_max_age():
     if lifetime is not None:
         return int(lifetime.total_seconds())
     return None
+
+
+def get_session_lifetime():
+    """Resolve the absolute lifetime of a session.
+
+    A session starts when the user authenticates and cannot be extended past this
+    limit by rotating the refresh token: once it is reached the user has to log in
+    again.  Defaults to ``DEFAULT_SESSION_LIFETIME``.
+
+    Setting ``JWT_ALLAUTH_SESSION_LIFETIME`` to ``None`` disables the limit, which
+    means a leaked refresh token can be renewed indefinitely.  This is discouraged.
+    """
+    lifetime = getattr(settings, "JWT_ALLAUTH_SESSION_LIFETIME", DEFAULT_SESSION_LIFETIME)
+    if lifetime is not None and not isinstance(lifetime, timedelta):
+        raise ImproperlyConfigured(
+            "jwt-allauth: JWT_ALLAUTH_SESSION_LIFETIME must be a datetime.timedelta or None."
+        )
+    return lifetime
 
 
 def _get_cookie_secure():